move VariableAssignment to operators.py
