+++ src/astx/variables.py
@@ -119,39 +119,6 @@ class InlineVariableDeclaration(Expr):
 
 @public
 @typechecked
-class VariableAssignment(StatementType):
-    """AST class for variable declaration."""
-
-    name: str
-    value: Expr
-
-    def __init__(
-        self,
-        name: str,
-        value: Expr,
-        loc: SourceLocation = NO_SOURCE_LOCATION,
-        parent: Optional[ASTNodes] = None,
-    ) -> None:
-        """Initialize the VarExprAST instance."""
-        super().__init__(loc=loc, parent=parent)
-        self.loc = loc
-        self.name = name
-        self.value = value
-        self.kind = ASTKind.VariableAssignmentKind
-
-    def __str__(self) -> str:
-        """Return a string that represents the object."""
-        return f"VariableAssignment[{self.name}]"
-
-    def get_struct(self, simplified: bool = False) -> ReprStruct:
-        """Return the AST structure of the object."""
-        key = str(self)
-        value = self.value.get_struct(simplified)
-        return self._prepare_struct(key, value, simplified)
-
-
-@public
-@typechecked
 class Variable(DataTypeOps):
     """AST class for the variable usage."""
